+++ deprecated/fabric-models-v0/build.gradle
@@ -1,0 +1,12 @@
+archivesBaseName = "fabric-models-v0"
+version = getSubprojectVersion(project)
+
+moduleDependencies(project, [
+	'fabric-api-base',
+	'fabric-model-loading-api-v1'
+])
+
+testDependencies(project, [
+	':fabric-rendering-v1',
+	':fabric-resource-loader-v0'
+])
