+++ deprecated/fabric-models-v0/src/client/resources/fabric.mod.json
@@ -1,0 +1,27 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-models-v0",
+  "name": "Fabric Models (v0)",
+  "version": "${version}",
+  "environment": "client",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-models-v0/icon.png",
+  "contact": {
+    "homepage": "https://fabricmc.net",
+    "irc": "irc://irc.esper.net:6667/fabric",
+    "issues": "https://github.com/FabricMC/fabric/issues",
+    "sources": "https://github.com/FabricMC/fabric"
+  },
+  "authors": [
+    "FabricMC"
+  ],
+  "depends": {
+    "fabricloader": ">=0.4.0",
+    "fabric-api-base": "*",
+    "fabric-model-loading-api-v1": "*"
+  },
+  "description": "Hooks for models and model loading.",
+  "custom": {
+    "fabric-api:module-lifecycle": "stable"
+  }
+}
