+++ fabric-model-loading-api-v1/src/client/resources/assets/fabric-model-loading-api-v1/icon.png
