+++ fabric-model-loading-api-v1/src/client/resources/fabric.mod.json
@@ -1,0 +1,35 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-model-loading-api-v1",
+  "name": "Fabric Model Loading API (v1)",
+  "version": "${version}",
+  "environment": "client",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-model-loading-api-v1/icon.png",
+  "contact": {
+    "homepage": "https://fabricmc.net",
+    "irc": "irc://irc.esper.net:6667/fabric",
+    "issues": "https://github.com/FabricMC/fabric/issues",
+    "sources": "https://github.com/FabricMC/fabric"
+  },
+  "authors": [
+    "FabricMC"
+  ],
+  "depends": {
+    "fabricloader": ">=0.14.21",
+    "fabric-api-base": "*"
+  },
+  "description": "Provides hooks for model loading.",
+  "mixins": [
+    {
+      "environment": "client",
+      "config": "fabric-model-loading-api-v1.mixins.json"
+    }
+  ],
+  "custom": {
+    "fabric-api:module-lifecycle": "stable",
+    "loom:injected_interfaces": {
+      "net/minecraft/class_1092": [ "net/fabricmc/fabric/api/client/model/loading/v1/FabricBakedModelManager" ]
+    }
+  }
+}
