--- fabric-models-v0/build.gradle
@@ -1,9 +1,0 @@
-archivesBaseName = "fabric-models-v0"
-version = getSubprojectVersion(project)
-
-moduleDependencies(project, ['fabric-api-base'])
-
-testDependencies(project, [
-	':fabric-rendering-v1',
-	':fabric-resource-loader-v0'
-])
