--- fabric-models-v0/src/client/resources/fabric-models-v0.mixins.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.client.model",
-  "compatibilityLevel": "JAVA_16",
-  "client": [
-    "BakedModelManagerMixin",
-    "ModelLoaderMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
