--- fabric-models-v0/src/testmodClient/resources/fabric.mod.json
@@ -1,17 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-models-v0-testmod",
-  "name": "Fabric Models (v0) Test Mod",
-  "version": "1.0.0",
-  "environment": "client",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-models-v0": "*",
-    "fabric-resource-loader-v0": "*"
-  },
-  "entrypoints": {
-    "client": [
-      "net.fabricmc.fabric.test.model.ModelTestModClient"
-    ]
-  }
-}
