+++ fabric-renderer-api-v1/src/testmodClient/java/net/fabricmc/fabric/test/renderer/simple/client/RendererClientTest.java
@@ -18,30 +18,51 @@ package net.fabricmc.fabric.test.renderer.simple.client;
 
 import static net.fabricmc.fabric.test.renderer.simple.RendererTest.id;
 
+import java.util.HashSet;
+import java.util.Set;
+
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.registry.Registries;
+import net.minecraft.util.Identifier;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
-import net.fabricmc.fabric.api.client.model.ModelLoadingRegistry;
+import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
 import net.fabricmc.fabric.test.renderer.simple.FrameBlock;
 import net.fabricmc.fabric.test.renderer.simple.RendererTest;
 
 public final class RendererClientTest implements ClientModInitializer {
+	private static final Set<Identifier> FRAME_MODELS = new HashSet<>();
+
 	@Override
 	public void onInitializeClient() {
-		ModelLoadingRegistry.INSTANCE.registerResourceProvider(manager -> new FrameModelResourceProvider());
-		ModelLoadingRegistry.INSTANCE.registerVariantProvider(manager -> new PillarModelVariantProvider());
-
 		for (FrameBlock frameBlock : RendererTest.FRAMES) {
 			// We don't specify a material for the frame mesh,
 			// so it will use the default material, i.e. the one from BlockRenderLayerMap.
 			BlockRenderLayerMap.INSTANCE.putBlock(frameBlock, RenderLayer.getCutoutMipped());
 
 			String itemPath = Registries.ITEM.getId(frameBlock.asItem()).getPath();
-			FrameModelResourceProvider.FRAME_MODELS.add(id("item/" + itemPath));
+			FRAME_MODELS.add(id("item/" + itemPath));
 		}
 
-		FrameModelResourceProvider.FRAME_MODELS.add(id("block/frame"));
+		FRAME_MODELS.add(id("block/frame"));
+
+		ModelLoadingPlugin.register(pluginContext -> {
+			pluginContext.resourceProviders().register((resourceId, context) -> {
+				if (FRAME_MODELS.contains(resourceId)) {
+					return new FrameUnbakedModel();
+				}
+
+				return null;
+			});
+
+			pluginContext.variantProviders().register((modelId, context) -> {
+				if (RendererTest.PILLAR_ID.equals(modelId)) {
+					return new PillarUnbakedModel();
+				}
+
+				return null;
+			});
+		});
 	}
 }
