+++ gradle.properties
@@ -38,6 +38,7 @@ fabric-loot-api-v2-version=1.1.38
 fabric-loot-tables-v1-version=1.1.42
 fabric-message-api-v1-version=5.1.6
 fabric-mining-level-api-v1-version=2.1.48
+fabric-model-loading-api-v1-version=1.0.0
 fabric-models-v0-version=0.3.35
 fabric-networking-api-v1-version=1.3.8
 fabric-networking-v0-version=0.3.48
