+++ settings.gradle
@@ -33,7 +33,7 @@ include 'fabric-lifecycle-events-v1'
 include 'fabric-loot-api-v2'
 include 'fabric-message-api-v1'
 include 'fabric-mining-level-api-v1'
-include 'fabric-models-v0'
+include 'fabric-model-loading-api-v1'
 include 'fabric-networking-api-v1'
 include 'fabric-object-builder-api-v1'
 include 'fabric-particles-v1'
@@ -61,6 +61,7 @@ include 'deprecated:fabric-command-api-v1'
 include 'deprecated:fabric-containers-v0'
 include 'deprecated:fabric-events-lifecycle-v0'
 include 'deprecated:fabric-keybindings-v0'
+include 'deprecated:fabric-models-v0'
 include 'deprecated:fabric-networking-v0'
 include 'deprecated:fabric-renderer-registries-v1'
 include 'deprecated:fabric-rendering-v0'
